config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

TOKEN = os.getenv("DISCORD_TOKEN")
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

# Load extra keys from comma-separated string
keys_str = os.getenv("GEMINI_API_KEYS", "")
GEMINI_API_KEYS = [k.strip() for k in keys_str.split(",") if k.strip()]

if not GEMINI_API_KEYS:
     GEMINI_API_KEYS = [GEMINI_API_KEY]

PLACEMENT_POINTS = {
    1: 12,
    2: 9,
    3: 8,
    4: 7,
    5: 6,
    6: 5,
    7: 4,
    8: 3,
    9: 2,
    10: 1,
    11: 0,
    12: 0
}

KILL_POINTS = 1

