requirements.txt
discord.py
python-dotenv
google-generativeai
supabase
pillow
pandas
